refactor: 패키지 구조 변경 및 의존성 역전, 실패하는 테스트 통과하도록 수정
