+++ src/main/java/shopping/application/CartService.java
@@ -3,18 +3,18 @@ package shopping.application;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import shopping.domain.cart.CartItem;
+import shopping.domain.cart.CartItemRepository;
 import shopping.domain.cart.CartItems;
 import shopping.domain.cart.Quantity;
 import shopping.domain.product.Product;
+import shopping.domain.product.ProductRepository;
 import shopping.domain.user.User;
+import shopping.domain.user.UserRepository;
 import shopping.dto.CartCreateRequest;
 import shopping.dto.CartResponse;
 import shopping.dto.QuantityUpdateRequest;
 import shopping.exception.ErrorType;
 import shopping.exception.ShoppingException;
-import shopping.repository.CartItemRepository;
-import shopping.repository.ProductRepository;
-import shopping.repository.UserRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
