+++ src/main/java/shopping/application/TokenProvider.java
@@ -1,0 +1,5 @@
+package shopping.application;
+
+public interface TokenProvider {
+    String create(final String payload);
+}
