+++ src/main/java/shopping/config/interceptor/TokenConsumer.java
@@ -1,0 +1,7 @@
+package shopping.config.interceptor;
+
+public interface TokenConsumer {
+    String getPayload(final String token);
+
+    boolean validateToken(final String token);
+}
