+++ src/main/java/shopping/domain/product/ProductRepository.java
@@ -1,8 +1,7 @@
-package shopping.repository;
+package shopping.domain.product;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import shopping.domain.product.Product;
 
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Long> {
