+++ src/main/java/shopping/exception/GlobalExceptionHandler.java
@@ -2,19 +2,20 @@ package shopping.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import shopping.dto.ErrorResponse;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
-    @ExceptionHandler(ShoppingException.class)
+    @ExceptionHandler({ShoppingException.class, HttpMessageNotReadableException.class})
     public ResponseEntity<ErrorResponse> handleShoppingException(final ShoppingException e) {
         return new ResponseEntity<>(new ErrorResponse(e.getMessage()), e.getStatus());
     }
-
+    
     @ExceptionHandler(Exception.class)
-    public ResponseEntity<ErrorResponse> handleException(final Exception e) {
+    public ResponseEntity<ErrorResponse> handleException(Exception e) {
         return new ResponseEntity<>(new ErrorResponse("Server Error"), HttpStatus.INTERNAL_SERVER_ERROR);
     }
 }
