--- src/main/java/shopping/infrastructure/TokenProvider.java
@@ -1,9 +1,0 @@
-package shopping.infrastructure;
-
-public interface TokenProvider {
-    String create(final String payload);
-
-    String getPayload(final String token);
-
-    boolean validateToken(final String token);
-}
