--- src/test/java/shopping/repository/ProductRepositoryTest.java
@@ -1,28 +1,0 @@
-package shopping.repository;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import shopping.domain.product.Product;
-
-import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-@DataJpaTest
-class ProductRepositoryTest {
-
-    @Autowired
-    ProductRepository productRepository;
-
-    @Test
-    @DisplayName("모든 상품을 조회한다.")
-    void findAll() {
-        // when
-        List<Product> products = productRepository.findAll();
-
-        // then
-        assertThat(products).hasSize(2);
-    }
-}
